start making some utility functions and clean up some yaml
